+++ Libplanet.Tests/Action/WorldTest.cs
@@ -1,4 +1,5 @@
 using System;
+using System.Collections.Generic;
 using System.Linq;
 using System.Numerics;
 using Libplanet.Action;
@@ -370,55 +371,23 @@ namespace Libplanet.Tests.Action
         }
 
         [Fact]
-        public virtual void SetValidator()
+        public void SetValidatorSet()
         {
-            Assert.Throws<ArgumentOutOfRangeException>(() =>
-                _initWorld.SetValidator(new Validator(new PrivateKey().PublicKey, -1))
-            );
+            const int newValidatorCount = 6;
+            var world = _initWorld;
+            var keys = Enumerable
+                .Range(0, newValidatorCount)
+                .Select(i => new PrivateKey())
+                .ToList();
+            var validatorSet = new ValidatorSet(
+                keys.Select(key => new Validator(key.PublicKey, 1)).ToList());
+            world = world.SetValidatorSet(validatorSet);
 
-            var initCount = _keys.Length;
-            var key3 = new PrivateKey().PublicKey;
-            var key4 = new PrivateKey().PublicKey;
+            Assert.Equal(newValidatorCount, world.GetValidatorSet().TotalCount);
+            Assert.NotEqual(_initWorld.GetValidatorSet(), world.GetValidatorSet());
 
-            IWorld delta = _initWorld;
-            // delta already has 3 validators
-            Assert.Equal(initCount, delta.GetValidatorSet().TotalCount);
-
-            // nothing happens trying to delete non existing validator
-            delta = delta.SetValidator(new Validator(key3, 0));
-            Assert.Equal(initCount, delta.GetValidatorSet().TotalCount);
-
-            // add key 3 to the validator set
-            delta = delta.SetValidator(new Validator(key3, 1));
-            Assert.Equal(initCount + 1, delta.GetValidatorSet().TotalCount);
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key3, 1)));
-            Assert.False(delta.GetValidatorSet().Contains(new Validator(key4, 1)));
-
-            // add key 4 to the validator set
-            delta = delta.SetValidator(new Validator(key4, 1));
-            Assert.Equal(initCount + 2, delta.GetValidatorSet().TotalCount);
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key3, 1)));
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key4, 1)));
-
-            // remove key 3 from the validator set
-            delta = delta.SetValidator(new Validator(key3, 0));
-            Assert.Equal(initCount + 1, delta.GetValidatorSet().TotalCount);
-            Assert.False(delta.GetValidatorSet().Contains(new Validator(key3, 1)));
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key4, 1)));
-
-            // re-add key 3 to the validator set
-            delta = delta.SetValidator(new Validator(key3, 1));
-            Assert.Equal(initCount + 2, delta.GetValidatorSet().TotalCount);
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key3, 1)));
-            Assert.True(delta.GetValidatorSet().Contains(new Validator(key4, 1)));
-
-            // remove all keys from the validator set
-            delta = _keys.Aggregate(
-                delta,
-                (current, key) => current.SetValidator(new Validator(key.PublicKey, 0)));
-            delta = delta.SetValidator(new Validator(key3, 0));
-            delta = delta.SetValidator(new Validator(key4, 0));
-            Assert.Equal(0, delta.GetValidatorSet().TotalCount);
+            world = world.SetValidatorSet(new ValidatorSet());
+            Assert.Equal(0, world.GetValidatorSet().TotalCount);
         }
 
         protected FungibleAssetValue Value(int currencyIndex, BigInteger quantity) =>